move ParseConfig helper to sdk/helper/hclutil
