+++ helper/profiles/config_test.go
@@ -6,11 +6,11 @@ import (
 	"testing"
 
 	"github.com/hashicorp/hcl/hcl/ast"
-	"github.com/openbao/openbao/helper/random"
+	"github.com/openbao/openbao/sdk/v2/helper/hclutil"
 )
 
 func parseBlockList(hclStr, blockName string) (*ast.ObjectList, error) {
-	file, err := random.ParseConfig([]byte(hclStr))
+	file, err := hclutil.ParseConfig([]byte(hclStr))
 	if err != nil {
 		return nil, fmt.Errorf("failed to parse HCL: %w", err)
 	}
@@ -134,7 +134,7 @@ request "op1" {
   data      = { key = "value" }
 }
 `
-	file, err := random.ParseConfig([]byte(hclStr))
+	file, err := hclutil.ParseConfig([]byte(hclStr))
 	if err != nil {
 		t.Fatalf("failed to parse HCL: %v", err)
 	}
@@ -165,7 +165,7 @@ request {
   operation = "read"
 }
 `
-	file, err := random.ParseConfig([]byte(hclStr))
+	file, err := hclutil.ParseConfig([]byte(hclStr))
 	if err != nil {
 		t.Fatalf("failed to parse HCL: %v", err)
 	}
