+++ sdk/helper/hclutil/hcl.go
@@ -5,9 +5,12 @@ package hclutil
 
 import (
 	"fmt"
+	"strings"
 
 	multierror "github.com/hashicorp/go-multierror"
 	"github.com/hashicorp/hcl/hcl/ast"
+	hclParser "github.com/hashicorp/hcl/hcl/parser"
+	jsonParser "github.com/hashicorp/hcl/json/parser"
 )
 
 // CheckHCLKeys checks whether the keys in the AST list contains any of the valid keys provided.
@@ -36,4 +39,25 @@ func CheckHCLKeys(node ast.Node, valid []string) error {
 	}
 
 	return result
+}
+
+// isJson determines if the input is JSON, i.e if it starts with '{'
+func isJson(data []byte) bool {
+	// Trim whitespace and check if it starts with '{'
+	trimmed := strings.TrimSpace(string(data))
+	return strings.HasPrefix(trimmed, "{")
+}
+
+// ParseConfig parses HCL or JSON configuration data into an AST.
+// It automatically detects the format and uses the appropriate parser.
+// For HCL format, it uses ParseDontErrorOnDuplicateKeys to maintain
+// backward compatibility with duplicate keys.
+func ParseConfig(data []byte) (*ast.File, error) {
+	if isJson(data) {
+		// JSON format
+		return jsonParser.Parse(data)
+	} else {
+		// HCL format
+		return hclParser.ParseDontErrorOnDuplicateKeys(data)
+	}
 }
