+++ vault/policy.go
@@ -17,7 +17,6 @@ import (
 	"github.com/mitchellh/copystructure"
 	"github.com/openbao/openbao/helper/identity"
 	"github.com/openbao/openbao/helper/namespace"
-	"github.com/openbao/openbao/helper/random"
 	"github.com/openbao/openbao/sdk/v2/helper/hclutil"
 	"github.com/openbao/openbao/sdk/v2/helper/identitytpl"
 	"github.com/openbao/openbao/sdk/v2/logical"
@@ -261,7 +260,7 @@ func ParseACLPolicy(ns *namespace.Namespace, rules string) (*Policy, error) {
 // templated policy.
 func parseACLPolicyWithTemplating(ns *namespace.Namespace, rules string, performTemplating bool, entity *identity.Entity, groups []*identity.Group) (*Policy, error) {
 	// Parse the rules
-	root, err := random.ParseConfig([]byte(rules))
+	root, err := hclutil.ParseConfig([]byte(rules))
 	if err != nil {
 		return nil, fmt.Errorf("failed to parse policy: %w", err)
 	}
